solving some trikcy danglin ptr issues
not properly solved yet, but at least we understand them now

flat_file::iterator need a pointer back to the flat_file. Not just for
the buffer but just to be able to load anything. I don't see a way
around that.

That pointer back to the flat_file can easily dangle when you
move_assign or move_construct the flat_file (such as during
std::vector<flat_file> reasizing or removing of elements.)

TODO going to try to keep track of which iterators have been issued
and keep their pointer "back to the mothership" updated.