make use of parallel STL optional
and off by default
user can select it if compiing from source for target platofm

reason was: the binary built on Debian11 is a different version to
Debian 12 and recent Ubuntu. Static version of ttb not easily
available. Given the benefit it small, and only for batch processes,
it makes sense to make this optional.