disbale redundant #include <execution>
when HIBP_WITH_PSTL=OFF (now he default), then

./build.sh -c gcc -b debug

was failing.

This is because above config var disables linking
tbb (that was the desidred effect, for cross platform
compatibility). It also #if s out the relevant
code (std::execution::par_unseq) in 2 places.

However on gcc (not clang) and in debug buildtype
only (not release) the "mere inclusion of the header" makes the linker
fail with "undefined symbol". Release mode removes the unused include.

so now we #if out the #include as well

sort of fair enough, but hard to find