ensure we notify on finished_dls in all cases
1. otherwise if queuemmgt thread is too fast processing the last message
and is already in msg_cv.wait(), and then requests thread calls
finished_dls, but this (prior to this commit) did not contain a
msg_cv.notify_one, then queuemgt thread will hang there forever and
when it is manually killed by the user the buffers are not flushed and
are missing the last few kB of data on disk.

2. also ensure that finished_dls() is called when exceptions are
throws in the requests thread. This was not out immediate issue, but
"failed network" testing shows this is certainly required to prevent
hanging.

both should be determistic and not dependent upong cv.wait_for()

we can get stuck,