beginning to implement ntlm hashes
- templatised hibp::pawned_pw and provide 2 aliases

- used those aliases throughout

- started using them in with search.

- which required templatising toc

- using explicit instantiations, since there are only 2 types, they
  are entirely known, and we want to retain compilation advantage of
  .cpp files

- not yet templatised hibp_server as that will likely heavily caascade
want to make search fully work first.

- need to decide on filename extension convention, and eventually,
prob after its all templatised, some header block for the file to make
sure its the correct one

- considered various types of type erasure and polymorphism for this,
but since the basic branch on which type to use, always starts with
the cli and is a simple if at the t0op level which never has to be
stored in a variable, then simple duck typing with templates and
explicit instantiations seems the easiest and best. slightly increase
binary size, but not too verbose, I think, simple maintenance.

- compiles and runs (with some logic errors I am sure)